refactor: IntroActivity와 MeetingInfoActivity 연결
